+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - Support for HTTP Header in load balancer rule set
+- Support for `oci_metering_computation_usage` resource and `oci_metering_computation_configuration` datasource
 
 ## 3.82.0 (June 24, 2020)